requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.2.0
ldap3==2.9.1